refactor: move filters out of LogContentProvider into their own class
