+++ src/extension/src/constants/constants.ts
@@ -23,5 +23,10 @@ export const KEYWORD_HIGHLIGHT_CONFIGURATION_SETTING_NAME = "keywords.highlights
  */
 export const WELCOME_MESSAGE_CONFIGURATION_SETTING_NAME = "showWelcomeMessage";
 
+/**
+ * The date of the epoch. Used for filtering out log entries that do not have a timestamp.
+ */
+export const EPOCH_DATE = new Date(0);
+
