+++ src/extension/src/providers/content/LogContentProvider.ts
@@ -15,6 +15,7 @@ import * as vscode from "vscode";
 
 import { ConfigurationManager } from "../../configuration/ConfigurationManager";
 import { DocumentLocationManager } from "../../configuration/DocumentLocationManager";
+import { EPOCH_DATE } from "../../constants/constants";
 import { ERROR_REGEX, GUID_REGEX, WARN_REGEX, WEB_DATE_REGEX } from "../../constants/regex";
 import { PostMessageDisposableService } from "../../service/PostMessageDisposableService";
 import { ScopedILogger } from "../../telemetry/ILogger";
@@ -22,8 +23,8 @@ import { ITelemetryLogger } from "../../telemetry/ITelemetryLogger";
 import { throwIfCancellation } from "../../utils/throwIfCancellation";
 
 import { HarFileProvider } from "./HarFileProvider";
+import { LogContentFilters } from "./LogContentFilters";
 import { LogEntry } from "./LogEntry";
-
 
 type ServiceFiles = {
     /**
@@ -48,82 +49,9 @@ const LOG_LEVEL_REGEX: Record<LogLevel, RegExp> = {
 };
 
 /**
- * The date of the epoch. Used for filtering out log entries that do not have a timestamp.
- */
-const EPOCH_DATE = new Date(0);
-
-/**
  * A content provider that transforms the content of a log file.
  */
 export class LogContentProvider extends PostMessageDisposableService implements vscode.TextDocumentContentProvider {
-    /**
-     * Filter out log entries that are before this date.
-     * This field is the string representation of a date.
-     */
-    private _timeFilterFrom: string | null = null;
-
-    /**
-     * Sets both {@link _timeFilterFrom} and {@link timeFilterFromDate}.
-     */
-    private set timeFilterFrom(value: string | null) {
-        this._timeFilterFrom = value;
-        if (this._timeFilterFrom && !isNaN(Date.parse(this._timeFilterFrom))) {
-            this.timeFilterFromDate = new Date(this._timeFilterFrom);
-        } else {
-            this.timeFilterFromDate = null;
-        }
-        this.postMessageService.sendAndForget({ command: "updateTimeFilters", data: { fromDate: value ?? null } });
-    }
-
-    /**
-     * Date representation of {@link _timeFilterFrom}.
-     */
-    private timeFilterFromDate: Date | null = null;
-
-    /**
-     * The minimum date that can be used as a filter except if the user wants to display all entries.
-     */
-    private minimumDate: string | null = new Date(1000).toISOString();
-
-    /**
-     * Filter out log entries that are after this date.
-     * This field is the string representation of a date.
-     */
-    private _timeFilterTill: string | null = null;
-
-    /**
-     * Sets both {@link _timeFilterTill} and {@link timeFilterTillDate}.
-     */
-    private set timeFilterTill(value: string | null) {
-        this._timeFilterTill = value;
-        if (this._timeFilterTill && !isNaN(Date.parse(this._timeFilterTill))) {
-            this.timeFilterTillDate = new Date(this._timeFilterTill);
-        } else {
-            this.timeFilterTillDate = null;
-        }
-        this.postMessageService.sendAndForget({ command: "updateTimeFilters", data: { tillDate: value ?? null } });
-    }
-
-    /**
-     * Date representation of {@link _timeFilterTill}.
-     */
-    private timeFilterTillDate: Date | null = null;
-
-    /**
-     * Filter out log entries that do not contain either of these keywords.
-     */
-    private keywordFilters: string[] = [];
-
-    /**
-     * The session id marks the starting point of the log entries to filter to.
-     */
-    private sessionId: string | null = null;
-
-    /**
-     * The currently disabled log levels.
-     */
-    private disabledLogLevels: LogLevel[] = [];
-
     /**
      * A list of log files that are used to render the text content.
      * This list will only be set on the first call to provideTextDocumentContent.
@@ -203,9 +131,9 @@ export class LogContentProvider extends PostMessageDisposableService implements 
     ];
 
     /**
-     * A list of functions that should be called after {@link provideTextDocumentContent} is finished.
+     * Class that manages the filters for the log content.
      */
-    private readonly filterMessagesToRespondTo: PostMessageEventRespondFunction[] = [];
+    private readonly filters: LogContentFilters;
 
     /**
      * A list of functions that should be called after {@link provideTextDocumentContent} is finished.
@@ -257,194 +185,30 @@ export class LogContentProvider extends PostMessageDisposableService implements 
      * @param logger The logger.
      */
     constructor(
-        private readonly onFilterChangeEvent: vscode.Event<TimeFilterChangedEvent>,
+        onFilterChangeEvent: vscode.Event<TimeFilterChangedEvent>,
         private readonly postMessageService: IPostMessageService,
-        private readonly configurationManager: ConfigurationManager,
+        configurationManager: ConfigurationManager,
         private readonly documentLocationManager: DocumentLocationManager,
         logger: ITelemetryLogger
     ) {
         super();
-        this.registerMessageHandlers();
-
-        // set the "timeFilterFrom" to a second after timestamp 0.
-        // This is done so that we ignore all events that do not have a timestamp.
-        this.timeFilterFrom = this.minimumDate;
         this.logger = logger.createLoggerScope("LogContentProvider");
+
+        this.filters = new LogContentFilters(
+            onFilterChangeEvent,
+            postMessageService,
+            configurationManager,
+            () => this.triggerDocumentChange(),
+            () => this.logEntryCache,
+            logger
+        );
+        this.registerDisplaySettingEvents();
+
         this.watcher = vscode.workspace.createFileSystemWatcher("**/*.{log,txt}", false, false, false);
         this.registerFileWatcherEvents();
 
-        this.setupKeywordFiltersFromConfiguration();
-        this.setupLogLevelFiltersFromConfiguration();
-        this.setupTimeFiltersFromConfiguration();
-
         this.harFileProvider = new HarFileProvider(logger);
         this.cursorPositionAfterContentGeneration = configurationManager.restoredCursorPosition;
-    }
-
-    /**
-     * Registers all post message handlers.
-     */
-    private registerMessageHandlers() {
-        this.registerFilterEvents();
-        this.registerDisplaySettingEvents();
-    }
-
-    /**
-     * Sets up the keyword filters from the configuration.
-     */
-    private setupKeywordFiltersFromConfiguration() {
-        this.keywordFilters = this.configurationManager.keywordFilters.filter(kw => kw.isChecked).map(kw => kw.keyword);
-        this.logger.info("setupKeywordFiltersFromConfiguration");
-        this.postMessageService.sendAndForget({ command: "updateNumberOfActiveFilters", data: this.getNumberOfActiveFilters() });
-        this.postMessageService.sendAndForget({
-            command: "setKeywordFiltersFromConfiguration",
-            data: this.configurationManager.keywordFilters.map(kw => ({ value: kw.keyword, isChecked: kw.isChecked })),
-        });
-    }
-
-    /**
-     * Sets up the log level filters from the configuration.
-     */
-    private setupLogLevelFiltersFromConfiguration() {
-        this.logger.info("setupLogLevelFiltersFromConfiguration");
-
-        this.disabledLogLevels = this.configurationManager.disabledLogLevels;
-
-        this.postMessageService.sendAndForget({ command: "updateNumberOfActiveFilters", data: this.getNumberOfActiveFilters() });
-        this.postMessageService.sendAndForget({
-            command: "setLogLevelFromConfiguration",
-            data: this.disabledLogLevels,
-        });
-    }
-
-    /**
-     * Sets up the time filters from the configuration.
-     */
-    private setupTimeFiltersFromConfiguration() {
-        this.logger.info("setupTimeFiltersFromConfiguration");
-
-        if (this.configurationManager.enabledTimeFilters.fromDate !== undefined) {
-            this.logger.info("setupTimeFiltersFromConfiguration.fromDate", undefined, {
-                fromDate: this.configurationManager.enabledTimeFilters.fromDate?.toString(),
-            });
-            this.timeFilterFrom = this.configurationManager.enabledTimeFilters.fromDate;
-        }
-
-        if (this.configurationManager.enabledTimeFilters.tillDate !== undefined) {
-            this.logger.info("setupTimeFiltersFromConfiguration.tillDate", undefined, {
-                tillDate: this.configurationManager.enabledTimeFilters.tillDate?.toString(),
-            });
-            this.timeFilterTill = this.configurationManager.enabledTimeFilters.tillDate;
-        }
-    }
-
-    /**
-     * Registers the filter events for the log content provider.
-     * This method is called in the constructor.
-     *
-     * Events are called when the user changes a filter through the webview.
-     * They are dispatched by the {@link ExtensionPostMessageService}.
-     */
-    private registerFilterEvents() {
-        const filterCheckboxStateChange = this.postMessageService.registerMessageHandler("filterCheckboxStateChange", (event, respond) => {
-            if (event.isChecked) {
-                this.keywordFilters.push(event.value);
-            } else {
-                this.keywordFilters = this.keywordFilters.filter(keyword => keyword !== event.value);
-            }
-            this.filterMessagesToRespondTo.push(respond);
-            this.triggerDocumentChange();
-        });
-        this.unregisterListeners.push(filterCheckboxStateChange);
-
-        const filterLogLevel = this.postMessageService.registerMessageHandler("filterLogLevel", (event, respond) => {
-            if (event.isChecked) {
-                this.disabledLogLevels = this.disabledLogLevels.filter(level => level !== event.logLevel);
-            } else {
-                if (!this.disabledLogLevels.includes(event.logLevel)) {
-                    this.disabledLogLevels.push(event.logLevel);
-                }
-            }
-            this.filterMessagesToRespondTo.push(respond);
-            this.triggerDocumentChange();
-        });
-        this.unregisterListeners.push(filterLogLevel);
-
-        const filterTime = this.postMessageService.registerMessageHandler("filterTime", (event, respond) => {
-            if (event.fromDate || event.fromDate === "") {
-                if (event.fromDate === "") {
-                    this.timeFilterFrom = this.minimumDate;
-                } else {
-                    this.timeFilterFrom = event.fromDate;
-                }
-            }
-
-            if (event.tillDate || event.tillDate === "") {
-                this.timeFilterTill = event.tillDate;
-            }
-            this.filterMessagesToRespondTo.push(respond);
-            this.triggerDocumentChange();
-        });
-        this.unregisterListeners.push(filterTime);
-
-        const filterRemoveEntriesWithNoEventTime = this.postMessageService.registerMessageHandler("filterNoEventTime", (event, respond) => {
-            if (event.removeEntriesWithNoEventTime === true) {
-                this.minimumDate = new Date(1000).toISOString();
-                this.timeFilterFrom = this.minimumDate;
-            } else if (event.removeEntriesWithNoEventTime === false) {
-                this.timeFilterFrom = null;
-                this.minimumDate = null;
-            }
-            this.filterMessagesToRespondTo.push(respond);
-            this.triggerDocumentChange();
-        });
-        this.unregisterListeners.push(filterRemoveEntriesWithNoEventTime);
-
-        const filterSessionId = this.postMessageService.registerMessageHandler("filterSessionId", (event, respond) => {
-            if (event.isChecked) {
-                this.sessionId = event.sessionId;
-                const sessionIdLogEntry = this.findEarliestSessionIdInLogEntries();
-                if (sessionIdLogEntry) {
-                    // subtract 1 second from the timestamp to make sure we include the log entry with the session id
-                    const filterFrom = new Date(sessionIdLogEntry.date.getTime() - 1000).toISOString();
-                    this.logger.info("filterSessionId.success", undefined, { sessionId: this.sessionId, filterFrom });
-                    this.timeFilterFrom = filterFrom;
-                } else {
-                    this.logger.logException(
-                        "filterSessionId.notFound",
-                        new Error(`Could not find log entry with session id: ${this.sessionId}`),
-                        undefined,
-                        {
-                            sessionId: this.sessionId,
-                        },
-                        true,
-                        "Session Id"
-                    );
-                }
-            } else {
-                this.sessionId = null;
-                this.timeFilterFrom = this.minimumDate;
-            }
-            this.filterMessagesToRespondTo.push(respond);
-            this.triggerDocumentChange();
-        });
-        this.unregisterListeners.push(filterSessionId);
-
-        this.onFilterChangeEvent(event => {
-            if (event.fromDate || event.fromDate === "") {
-                if (event.fromDate === "") {
-                    this.timeFilterFrom = this.minimumDate;
-                } else {
-                    this.timeFilterFrom = event.fromDate;
-                }
-            }
-
-            if (event.tillDate || event.tillDate === "") {
-                this.timeFilterTill = event.tillDate;
-            }
-
-            this.triggerDocumentChange();
-        });
     }
 
     /**
@@ -503,6 +267,7 @@ export class LogContentProvider extends PostMessageDisposableService implements 
     public override dispose() {
         super.dispose();
         this.watcher.dispose();
+        this.filters.dispose();
     }
 
     /**
@@ -534,12 +299,7 @@ export class LogContentProvider extends PostMessageDisposableService implements 
         this.groupedLogEntries = new Map();
         this.harFileProvider.clearCache();
         if (resetFilters) {
-            this.keywordFilters = [];
-            this.disabledLogLevels = [];
-            this.sessionId = null;
-            this.timeFilterFrom = this.minimumDate;
-            this.timeFilterTill = null;
-            this.postMessageService.sendAndForget({ command: "updateNumberOfActiveFilters", data: 0 });
+            this.filters.reset();
         }
 
         this.changeTrigger++;
@@ -552,74 +312,6 @@ export class LogContentProvider extends PostMessageDisposableService implements 
     private triggerDocumentChange() {
         this.changeTrigger++;
         this._onDidChange.fire(LogContentProvider.documentUri);
-    }
-
-    /**
-     * Tries to find a log entry that contains the session id and returns the entry.
-     * There might be multiple log entries with the same session id, but we need to find the earliest one.
-     * @returns The first log entry that contains the session id.
-     */
-    private findEarliestSessionIdInLogEntries(): LogEntry | null {
-        if (!this.sessionId) {
-            return null;
-        }
-
-        let foundEntry: LogEntry | null = null;
-        for (const logEntry of this.logEntryCache) {
-            if (logEntry.text.includes(this.sessionId)) {
-                // make sure it's not the summary.txt file
-                if (logEntry.service === "summary") {
-                    this.logger.info("findEarliestSessionIdInLogEntries.foundInSummary", undefined, { sessionId: this.sessionId });
-                    continue;
-                }
-
-                // make sure the entry has a timestamp (not the epoch date)
-                if (logEntry.date.getTime() === EPOCH_DATE.getTime()) {
-                    this.logger.info(
-                        "findEarliestSessionIdInLogEntries.noDate",
-                        `Found log entry with session id: ${logEntry.text} but it has no timestamp.`,
-                        { sessionId: this.sessionId }
-                    );
-                    continue;
-                }
-
-                if (!foundEntry) {
-                    this.logger.info("findEarliestSessionIdInLogEntries.foundFirst", undefined, {
-                        sessionId: this.sessionId,
-                        service: logEntry.service,
-                    });
-                    foundEntry = logEntry;
-                } else {
-                    // if we found another log entry with the same session id, check if it's earlier
-                    if (logEntry.date.getTime() < foundEntry.date.getTime()) {
-                        this.logger.info("findEarliestSessionIdInLogEntries.foundEarlier", undefined, {
-                            sessionId: this.sessionId,
-                            service: logEntry.service,
-                        });
-                        foundEntry = logEntry;
-                    } else {
-                        this.logger.info("findEarliestSessionIdInLogEntries.foundLater", undefined, {
-                            sessionId: this.sessionId,
-                            service: logEntry.service,
-                        });
-                    }
-                }
-            }
-        }
-
-        return foundEntry;
-    }
-
-    /**
-     * Calculates the number of filters that are currently active.
-     * @returns The number of filters that are currently active.
-     */
-    public getNumberOfActiveFilters(): number {
-        const numberOfKeywordFilters = this.keywordFilters.length;
-        const numberOfTimeFilters = (this._timeFilterFrom ? 1 : 0) + (this._timeFilterTill ? 1 : 0);
-        const logLevelFilter = this.disabledLogLevels.length;
-
-        return numberOfKeywordFilters + numberOfTimeFilters + logLevelFilter;
     }
 
     /**
@@ -742,7 +434,7 @@ export class LogContentProvider extends PostMessageDisposableService implements 
                 // update the cursor position if we have a position to update
                 // we have to wait for the content to be generated before we can update the cursor position
                 this.updateCursorPosition(token);
-                
+
                 this.updateFileList(logEntries, filteredLogEntires, token);
 
                 return content;
@@ -774,12 +466,16 @@ export class LogContentProvider extends PostMessageDisposableService implements 
      * @param filteredLogEntries Map of the log entries after filtering and grouping. The key is the timestamp in seconds and the value is the log entries for that second.
      * @param token The cancellation token.
      */
-    private updateFileList(nonFilteredLogEntries: LogEntry[], filteredLogEntries: Map<number, LogEntry[]>, token: vscode.CancellationToken): void {
+    private updateFileList(
+        nonFilteredLogEntries: LogEntry[],
+        filteredLogEntries: Map<number, LogEntry[]>,
+        token: vscode.CancellationToken
+    ): void {
         this.logger.info("updateFileList");
         throwIfCancellation(token);
         const files: Record<string, LogFile> = {};
         const fileKeys: string[] = [];
-        
+
         for (const logEntry of nonFilteredLogEntries) {
             const serviceName = logEntry.service;
             if (!serviceName) {
@@ -831,15 +527,7 @@ export class LogContentProvider extends PostMessageDisposableService implements 
      * Responds to the messages that are waiting for the content to be generated.
      */
     private respondToMessages() {
-        const activeFilters = this.getNumberOfActiveFilters();
-        // pop all the filter messages and respond to them
-        while (this.filterMessagesToRespondTo.length > 0) {
-            const respond = this.filterMessagesToRespondTo.pop();
-            if (respond) {
-                this.logger.info("respondToMessages");
-                respond({ command: "updateNumberOfActiveFilters", data: activeFilters });
-            }
-        }
+        this.filters.respondToMessages();
 
         // pop all the display settings messages and respond to them
         while (this.displaySettingsToRespondTo.length > 0) {
@@ -1141,9 +829,9 @@ export class LogContentProvider extends PostMessageDisposableService implements 
      * @returns True if the log entry should stay, false if it should be filtered out.
      */
     private matchesKeywordFilter(logEntryLine: string): boolean {
-        if (this.keywordFilters.length > 0) {
+        if (this.filters.keywordFilters.length > 0) {
             // Check if the log entry contains any of the keywords
-            const keywordRegex = new RegExp(this.keywordFilters.join("|"));
+            const keywordRegex = new RegExp(this.filters.keywordFilters.join("|"));
             const keywordMatch = logEntryLine.match(keywordRegex);
             if (!keywordMatch) {
                 return false;
@@ -1158,12 +846,12 @@ export class LogContentProvider extends PostMessageDisposableService implements 
      * @returns A new regex to remove log levels.
      */
     private buildLogLevelRemoveRegEx = (): RegExp | null => {
-        if (this.disabledLogLevels.length === 0) {
+        if (this.filters.disabledLogLevels.length === 0) {
             return null;
         }
 
         const regExToUse: string[] = [];
-        for (const levelToRemove of this.disabledLogLevels) {
+        for (const levelToRemove of this.filters.disabledLogLevels) {
             regExToUse.push(LOG_LEVEL_REGEX[levelToRemove].source);
         }
         return new RegExp(regExToUse.join("|"));
@@ -1193,11 +881,11 @@ export class LogContentProvider extends PostMessageDisposableService implements 
      * @returns True if the group should stay, false if it should be filtered out.
      */
     private matchesTimeFilter(groupTimestamp: number): boolean {
-        if (this.timeFilterFromDate && groupTimestamp < this.timeFilterFromDate.getTime()) {
+        if (this.filters.timeFilterFromDate && groupTimestamp < this.filters.timeFilterFromDate.getTime()) {
             return false;
         }
 
-        if (this.timeFilterTillDate && groupTimestamp > this.timeFilterTillDate.getTime()) {
+        if (this.filters.timeFilterTillDate && groupTimestamp > this.filters.timeFilterTillDate.getTime()) {
             return false;
         }
 
@@ -1356,19 +1044,19 @@ export class LogContentProvider extends PostMessageDisposableService implements 
      */
     private getFileTypeFromService(service: string): LogFileType {
         switch (service) {
-                    case "Launcher":
-                    case "MSTeams":
-                    case "TeamsNotificationCenter":
-                    case "TeamsRespawnService":
-                    case "TeamsSwitcher":
-                    case "skylib":
-                    case "tscalling":
-                        return "desktop";
-                    case "HAR":
-                        return "har";
-                    default:
-                        return "web";
-                }
+            case "Launcher":
+            case "MSTeams":
+            case "TeamsNotificationCenter":
+            case "TeamsRespawnService":
+            case "TeamsSwitcher":
+            case "skylib":
+            case "tscalling":
+                return "desktop";
+            case "HAR":
+                return "har";
+            default:
+                return "web";
+        }
     }
 }
